simplify dependency graph for dependents of LoRA utilities
