+++ vllm/v1/cudagraph_dispatcher.py
@@ -7,7 +7,7 @@ from itertools import product
 from vllm.config import CUDAGraphMode, VllmConfig
 from vllm.forward_context import BatchDescriptor
 from vllm.logger import init_logger
-from vllm.lora.utils import get_captured_lora_counts
+from vllm.lora.shared import get_captured_lora_counts
 
 logger = init_logger(__name__)
